requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
deepface==0.0.75
tensorflow==2.13.0
scikit-learn==1.3.0
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2
sqlalchemy==2.0.20
